* libbase/tu_timer.cpp: Fixed a compiler warning. 	* server/asobj/NetStreamFfmpeg.{h,cpp}: Changed the way the decoding 	  loops works. We now use condition variables to pause the decoding 	  loop. We now disconnect from the soundhandler when pausing.

+++ libbase/tu_timer.cpp
@@ -61,7 +61,7 @@ double	tu_timer::profile_ticks_to_seconds(uint64_t ticks)
 
 uint64_t tu_timer::get_ticks()
 {
-	return get_profile_ticks()/1000.0;
+	return static_cast<uint64_t>(get_profile_ticks()/1000.0);
 }
